Вынести генерацию карточки и её данные в модуль
